+++ programs/quarry-mint-wrapper/src/lib.rs
@@ -13,8 +13,10 @@ use vipers::prelude::*;
 
 mod account_validators;
 mod instructions;
+mod state;
 
 use instructions::*;
+pub use state::*;
 
 declare_id!("QMWoBmAyJLAsA1Lh9ugMTw2gciTihncciphzdNzdZYV");
 
@@ -82,9 +84,9 @@ pub mod quarry_mint_wrapper {
         Ok(())
     }
 
-    /// --------------------------------
-    /// [Minter] instructions
-    /// --------------------------------
+    // --------------------------------
+    // [Minter] instructions
+    // --------------------------------
 
     /// Creates a new [Minter].
     #[deprecated(since = "5.0.0", note = "Use `new_minter_v2` instead.")]
@@ -232,100 +234,25 @@ pub struct PerformMint<'info> {
     pub token_program: Program<'info, Token>,
 }
 
-/// --------------------------------
-/// Account structs
-/// --------------------------------
+// --------------------------------
+// Account structs
+// --------------------------------
 
+/// Only an admin is allowed to use instructions containing this struct.
 #[derive(Accounts)]
 pub struct OnlyAdmin<'info> {
     /// The [MintWrapper].
-    #[account(mut)]
+    #[account(mut, has_one = admin @ ErrorCode::Unauthorized)]
     pub mint_wrapper: Account<'info, MintWrapper>,
     /// [MintWrapper::admin].
     pub admin: Signer<'info>,
 }
 
-/// --------------------------------
-/// PDA structs
-/// --------------------------------
+// --------------------------------
+// Events
+// --------------------------------
 
-/// Mint wrapper
-///
-/// ```ignore
-/// seeds = [
-///     b"MintWrapper",
-///     base.key().to_bytes().as_ref(),
-///     &[bump]
-/// ],
-///
-#[account]
-#[derive(Default)]
-pub struct MintWrapper {
-    /// Base account.
-    pub base: Pubkey,
-    /// Bump for allowing the proxy mint authority to sign.
-    pub bump: u8,
-    /// Maximum number of tokens that can be issued.
-    pub hard_cap: u64,
-
-    /// Admin account.
-    pub admin: Pubkey,
-    /// Next admin account.
-    pub pending_admin: Pubkey,
-
-    /// Mint of the token.
-    pub token_mint: Pubkey,
-    /// Number of [Minter]s.
-    pub num_minters: u64,
-
-    /// Total allowance outstanding.
-    pub total_allowance: u64,
-    /// Total amount of tokens minted through the [MintWrapper].
-    pub total_minted: u64,
-}
-
-impl MintWrapper {
-    pub const LEN: usize = 32 + 1 + 8 + 32 + 32 + 32 + 8 + 8 + 8;
-}
-
-/// One who can mint.
-///
-/// ```ignore
-/// seeds = [
-///     b"MintWrapperMinter",
-///     auth.mint_wrapper.key().to_bytes().as_ref(),
-///     minter_authority.key().to_bytes().as_ref(),
-///     &[bump]
-/// ],
-/// ```
-#[account]
-#[derive(Default)]
-pub struct Minter {
-    /// The mint wrapper.
-    pub mint_wrapper: Pubkey,
-    /// Address that can mint.
-    pub minter_authority: Pubkey,
-    /// Bump seed.
-    pub bump: u8,
-
-    /// Auto-incrementing index of the [Minter].
-    pub index: u64,
-
-    /// Limit of number of tokens that this [Minter] can mint.
-    pub allowance: u64,
-    /// Cumulative sum of the number of tokens ever minted by this [Minter].
-    pub total_minted: u64,
-}
-
-impl Minter {
-    pub const LEN: usize = 32 + 32 + 1 + 8 + 8 + 8;
-}
-
-/// --------------------------------
-/// Events
-/// --------------------------------
-
-/// Triggered when a [MintWrapper] is created.
+/// Emitted when a [MintWrapper] is created.
 #[event]
 pub struct NewMintWrapperEvent {
     /// The [MintWrapper].
@@ -340,7 +267,7 @@ pub struct NewMintWrapperEvent {
     pub token_mint: Pubkey,
 }
 
-/// Triggered when a [MintWrapper]'s admin is proposed.
+/// Emitted when a [MintWrapper]'s admin is proposed.
 #[event]
 pub struct MintWrapperAdminProposeEvent {
     /// The [MintWrapper].
@@ -353,7 +280,7 @@ pub struct MintWrapperAdminProposeEvent {
     pub pending_admin: Pubkey,
 }
 
-/// Triggered when a [MintWrapper]'s admin is transferred.
+/// Emitted when a [MintWrapper]'s admin is transferred.
 #[event]
 pub struct MintWrapperAdminUpdateEvent {
     /// The [MintWrapper].
@@ -366,7 +293,7 @@ pub struct MintWrapperAdminUpdateEvent {
     pub admin: Pubkey,
 }
 
-/// Triggered when a [Minter] is created.
+/// Emitted when a [Minter] is created.
 #[event]
 pub struct NewMinterEvent {
     /// The [MintWrapper].
@@ -382,7 +309,7 @@ pub struct NewMinterEvent {
     pub minter_authority: Pubkey,
 }
 
-/// Triggered when a [Minter]'s allowance is updated.
+/// Emitted when a [Minter]'s allowance is updated.
 #[event]
 pub struct MinterAllowanceUpdateEvent {
     /// The [MintWrapper].
@@ -398,7 +325,7 @@ pub struct MinterAllowanceUpdateEvent {
     pub allowance: u64,
 }
 
-/// Triggered when a [Minter] performs a mint.
+/// Emitted when a [Minter] performs a mint.
 #[event]
 pub struct MinterMintEvent {
     /// The [MintWrapper].
@@ -414,10 +341,7 @@ pub struct MinterMintEvent {
     pub destination: Pubkey,
 }
 
-/// --------------------------------
 /// Error Codes
-/// --------------------------------
-
 #[error_code]
 pub enum ErrorCode {
     #[msg("You are not authorized to perform this action.")]
